non auth can't use api method to add and delete

+++ app/api/topics/route.js
@@ -1,11 +1,23 @@
 import connect from "@/lib/databaseconnect";
+import { auth } from "@clerk/nextjs/server";
 import Topic from "@/models/topic";
 import { NextResponse } from "next/server";
 export async function POST(request) {
+  const { userId } = await auth();
+  if (!userId) {
+    return NextResponse.json(
+      { error: "Kamu bukan administrator 👉👈" },
+      { status: 401 },
+    );
+  }
   const { title, description } = await request.json();
   await connect();
   await Topic.create({ title, description });
-  return NextResponse.json({ message: "Topic dibuat" }, { status: 201 });
+  return NextResponse.json(
+    { userId },
+    { message: "Topic dibuat" },
+    { status: 201 },
+  );
 }
 
 export async function GET() {
@@ -15,8 +27,19 @@ export async function GET() {
 }
 
 export async function DELETE(request) {
+  const { userId } = await auth();
+  if (!userId) {
+    return NextResponse.json(
+      { error: "Kamu bukan administrator 👉👈" },
+      { status: 401 },
+    );
+  }
   const id = request.nextUrl.searchParams.get("id");
   await connect();
   await Topic.findByIdAndDelete(id);
-  return NextResponse.json({ message: "Topic dihapus" }, { status: 200 });
+  return NextResponse.json(
+    { userId },
+    { message: "Topic dihapus" },
+    { status: 200 },
+  );
 }
